Decompression updates
This prepares the addition of LZO decompression.
Some decompression methods need to be given the size of the destination
buffer. As a result _read_block() and the methods that use it
received a few changes to have them provide this size.
For uncompress(), right now 'size' is unused, and 'outsize' will only
be used for LZO.

Change the name attribute of the zlib compressor from zlib to gzip
to be consistent with the name given in compressor.c.